requirements.txt
asgiref==3.4.1
Django==3.2.5
django-summernote==0.8.11.6
django-tinymce==3.3.0
Pillow==8.3.1
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
